pyproject.toml
[build-system]
requires = ["setuptools >= 75.2.0"]
build-backend = "setuptools.build_meta"

[project]
name = "offender"
version = "0.0.1"
description = "Offender: Your friendly elf Offset Finder"
readme = "README.md"
license = { text = "MIT License" }
requires-python = ">= 3.12"

dependencies = [
    "sqlalchemy >= 2.0.36",
    "pydantic>=2.9.2"
]

classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
]


[project.optional-dependencies]
test = ["pytest", "pytest-benchmark"]
elftools_searcher = ["pyelftools~=0.31", "r2pipe"]
r2searcher = ["r2pipe~=1.9.4"]
linux_searcher = ["keystone-engine~=0.9.2"]
all = ["offender[elftools_searcher, r2searcher]"]
dev = ["offender[all, test]"]